fix(kernels): repo-wide int64 pointer offsets + fabric-free module benches
int64 pointer safety (the headline fix):
- Promote tl.program_id() (and python loop-index x stride products) to
  tl.int64 across all raw-indexing triton kernels. Fixes aug_atom TRAINING
  CUDA illegal-memory-access at L>=768: the backward dq-expand split offsets
  (pid*stride_dq_split, s*stride_split) overflowed int32 for atom-sized
  A*B. compute_efficient now runs to L=1024 AND beats memory_efficient, so
  it stays the default.
- Exception: make_block_ptr kernels (triangle_attention, baseline_dtv1) keep
  program_id(0)/start_m as int32 - triton hard-requires 32-bit block-ptr
  offsets/block_shape; only the int64 batch/head BASE offsets matter for
  overflow and those were already int64. CUTE int32 (tmem addr, RNG seed)
  left as-is (hardware/semantic, not overflowing pointers).

Benchmarks measure the raw module, not the framework:
- Replace Lightning Fabric with a no-op shim in bench.py module benches
  (setup_module->identity, backward->tensor.backward). Fabric added ~110us/
  step of GPU casts/copies that inflated latency and compressed speedups;
  models are already device+dtype placed. transition d=128 train 0.74->0.63ms.

Also folded in (same uncommitted worktree):
- transition sm100 dispatch: d=256/512 route to cute b2b_fwd + gatebwd_sm100
  (recovers gate-backward, faster than old_triton); d=128 stays triton b2b.
- augmented_attention: dispatch wrapper exposing compute_efficient kwarg
  (default True), used by the module.
- bias_only_attention: fix gate-calibration do_bench return unpacking +
  cache the static fallback so eager stops re-probing every call.
- cute probes: memoized get_max_active_clusters (was JIT-recompiling ~35ms).
- plot_csv: default figures to plots/<gpu>/ (raw CSVs stay in artifacts/).

Validated fabric-free on B200 (sm100), all 9 modules, L up to 1024:
0 IMA, 0 correctness failures; int64 vs int32 speed delta within +/-3% noise.

+++ src/miniworld_kernels/kernels/adaln/triton/fused3.py
@@ -21,7 +21,7 @@ _LN_CFG = [triton.Config({"BLOCK_M": bm}, num_warps=nw, num_stages=ns)
 @triton.jit
 def _ln_kernel(X, Y, W, M, N, eps, sx0, sx1, sy0, sy1,
               HAS_W: tl.constexpr, BLOCK_M: tl.constexpr, BLOCK_N: tl.constexpr, DT: tl.constexpr):
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     rm = row * BLOCK_M + tl.arange(0, BLOCK_M)
     rmask = rm < M
     cols = tl.arange(0, BLOCK_N)
@@ -68,7 +68,7 @@ def _gemm_gate_kernel(
     DT: tl.constexpr, BLOCK_M: tl.constexpr, BLOCK_N: tl.constexpr, BLOCK_K: tl.constexpr,
     GROUP_M: tl.constexpr,
 ):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     num_pid_m = tl.cdiv(M, BLOCK_M)
     num_pid_n = tl.cdiv(N, BLOCK_N)
     num_pid_in_group = GROUP_M * num_pid_n
@@ -129,7 +129,7 @@ def _gemm_gate_train_kernel(
     DT: tl.constexpr, BLOCK_M: tl.constexpr, BLOCK_N: tl.constexpr, BLOCK_K: tl.constexpr,
     GROUP_M: tl.constexpr,
 ):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     num_pid_m = tl.cdiv(M, BLOCK_M)
     num_pid_n = tl.cdiv(N, BLOCK_N)
     num_pid_in_group = GROUP_M * num_pid_n
@@ -184,7 +184,7 @@ def _gemm_gate_train(x_norm, cond_norm, Ws, Wb, scale_b):
 def _bwd_elem_kernel(DY, Xn, Gate, Dscale, Dxn, M, N,
                      sy0, sy1, sxn0, sxn1, sg0, sg1, sds0, sds1, sdx0, sdx1,
                      BLOCK_M: tl.constexpr, BLOCK_N: tl.constexpr, DT: tl.constexpr):
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     rm = row * BLOCK_M + tl.arange(0, BLOCK_M)
     rmask = rm < M
     cols = tl.arange(0, BLOCK_N)
