+++ src/miniworld_kernels/kernels/augmented_attention/triton/main.py
@@ -143,9 +143,9 @@ def _attn_fwd(
     BLOCK_D: tl.constexpr,
     GROUP_N: tl.constexpr,
 ):
-    start_m = tl.program_id(0)
+    start_m = tl.program_id(0).to(tl.int64)
     off_hz = tl.program_id(1).to(tl.int64)
-    start_d = tl.program_id(2)
+    start_d = tl.program_id(2).to(tl.int64)
     off_z = off_hz // H
     off_h = off_hz % H
     off_a = off_z // B
@@ -240,7 +240,7 @@ def _attn_bwd_preprocess(
     BLOCK_D: tl.constexpr,
     GROUP_N: tl.constexpr,
 ):
-    off_m = tl.program_id(0) * BLOCK_M + tl.arange(0, BLOCK_M)
+    off_m = tl.program_id(0).to(tl.int64) * BLOCK_M + tl.arange(0, BLOCK_M)
     off_z = tl.program_id(1).to(tl.int64)
     off_h = tl.program_id(2).to(tl.int64)
 
@@ -385,7 +385,7 @@ def _attn_bwd(
     BLOCK_D: tl.constexpr,
     GROUP_N: tl.constexpr,
 ):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     aid = tl.program_id(1).to(tl.int64)
     bhid = tl.program_id(2).to(tl.int64)
     bid = bhid // H
