+++ src/miniworld_kernels/kernels/gated_projection/triton/main.py
@@ -43,7 +43,7 @@ def sigmoid_gate_fwd_kernel(
     BLOCK_N: tl.constexpr,
     GROUP_M: tl.constexpr,
 ):
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     offset_row = tl.arange(0, BLOCK_M) + row * BLOCK_M
     offset_col = tl.arange(0, BLOCK_N)
     row_mask = offset_row < n_elements
@@ -76,7 +76,7 @@ def sigmoid_gate_bwd_kernel(
     BLOCK_N: tl.constexpr,
     GROUP_M: tl.constexpr,
 ):
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     offset_row = tl.arange(0, BLOCK_M) + row * BLOCK_M
     offset_col = tl.arange(0, BLOCK_N)
     row_mask = offset_row < n_elements
